Add files via upload
My client Artemis Financial is a consulting firm that specializes in individualized financial models for its clients. They use a web application programming interface (API). This includes multiple focuses in the financial such as savings, investing, insurance, and retirement. Offering multiple features for its users meant that the security level needed to increase and modernize its custom software. 

Solution: Find the company's main source of vulnerability, develop a Vulnerability Assessment Report, and establish an actionable plan to protect it from future threats.

- Additionally: Showing the value modern capacity of software security can offer advantages to a company's overall well - being.
This method is established by addressing current generation Cryptographic Hash functioning, and next generation encryption potential. One of the challenges overcame when generating a vulnerability report is researching a potential threat and validating methods used to remove it from the software. Moreover, finding positives and suppressing them in the report to generate a focused report for (high, and critical) bugs.

In order to validate the success of the software build application functionality is documented in the report.

The reason this skill of Software Security is so important because it uses a crucial process to protect companies that face today's vulnerabilities.


Tools, or coding practices that can be used for future task.

- Hash Function Encryption
- JUnit Testing
- Generating Vulnerability Reports
- Suppressing False Positive
- Filter Bug Focus Report (High, Critical)